Upgrade cypress to version 12
Add cypress task to obtain next none working day for use with integration tests. Moment object gets parsed as text when called within the test file.
