+++ cypress.config.ts
@@ -1,0 +1,130 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { defineConfig } from 'cypress'
+import 'reflect-metadata'
+import moment, { Moment } from 'moment/moment'
+import { resetStubs, verifyEndpointCalled } from './integration_tests/wiremock'
+import auth from './integration_tests/mockApis/auth'
+import tokenVerification from './integration_tests/mockApis/tokenVerification'
+import licence from './integration_tests/mockApis/licence'
+import community from './integration_tests/mockApis/community'
+import prisonerSearch from './integration_tests/mockApis/prisonerSearch'
+import prison from './integration_tests/mockApis/prison'
+import probationSearch from './integration_tests/mockApis/probationSearch'
+import events from './integration_tests/support/events'
+import UkBankHolidayFeedService from './server/services/ukBankHolidayFeedService'
+import { isBankHolidayOrWeekend } from './server/utils/utils'
+
+export default defineConfig({
+  chromeWebSecurity: false,
+  fixturesFolder: 'integration_tests/fixtures',
+  screenshotsFolder: 'integration_tests/screenshots',
+  videosFolder: 'integration_tests/videos',
+  reporter: 'cypress-multi-reporters',
+  reporterOptions: {
+    configFile: 'reporter-config.json',
+  },
+  videoUploadOnPasses: false,
+  taskTimeout: 60000,
+  viewportHeight: 1200,
+  viewportWidth: 1300,
+  e2e: {
+    setupNodeEvents(on) {
+      on('task', {
+        reset: resetStubs,
+        verifyEndpointCalled,
+
+        getSignInUrl: auth.getSignInUrl,
+        stubPrisonSignIn: auth.stubPrisonSignIn,
+        stubProbationSignIn: auth.stubProbationSignIn,
+        stubProbationAcoSignIn: auth.stubProbationAcoSignIn,
+        stubAuthUser: auth.stubUser,
+        stubAuthPing: auth.stubPing,
+        stubSystemToken: auth.systemToken,
+
+        stubTokenVerificationPing: tokenVerification.stubPing,
+
+        stubUpdateResponsibleCom: licence.stubUpdateResponsibleCom,
+        stubUpdateProbationTeam: licence.stubUpdateProbationTeam,
+        stubGetLicence: licence.stubGetLicence,
+        stubPostLicence: licence.stubPostLicence,
+        stubGetExistingLicenceForOffenderWithResult: licence.stubGetExistingLicenceForOffenderWithResult,
+        stubGetLicencesForOffender: licence.stubGetLicencesForOffender,
+        stubPutAppointmentPerson: licence.stubPutAppointmentPerson,
+        stubPutAppointmentTime: licence.stubPutAppointmentTime,
+        stubPutAppointmentAddress: licence.stubPutAppointmentAddress,
+        stubPutContactNumber: licence.stubPutContactNumber,
+        stubPutBespokeConditions: licence.stubPutBespokeConditions,
+        stubPutAdditionalConditions: licence.stubPutAdditionalConditions,
+        stubGetLicenceWithConditionToComplete: licence.stubGetLicenceWithConditionToComplete,
+        stubGetLicenceWithPssConditionToComplete: licence.stubGetLicenceWithPssConditionToComplete,
+        stubPutAdditionalConditionData: licence.stubPutAdditionalConditionData,
+        stubGetExistingLicencesForOffenders: licence.stubGetExistingLicencesForOffenders,
+        stubGetExistingLicenceForOffenderNoResult: licence.stubGetExistingLicenceForOffenderNoResult,
+        stubSubmitStatus: licence.stubSubmitStatus,
+        stubUpdateLicenceStatus: licence.stubUpdateLicenceStatus,
+        stubGetLicencesForStatus: licence.stubGetLicencesForStatus,
+        stubGetCompletedLicence: licence.stubGetCompletedLicence,
+        stubRecordAuditEvent: licence.stubRecordAuditEvent,
+        stubCreateVariation: licence.stubCreateVariation,
+        stubUpdateSpoDiscussion: licence.stubUpdateSpoDiscussion,
+        stubUpdateVloDiscussion: licence.stubUpdateVloDiscussion,
+        stubUpdateReasonForVariation: licence.stubUpdateReasonForVariation,
+        stubGetLicenceVariationInProgress: licence.stubGetLicenceVariationInProgress,
+        stubDiscardLicence: licence.stubDiscardLicence,
+        stubUpdatePrisonInformation: licence.stubUpdatePrisonInformation,
+        stubUpdateSentenceDates: licence.stubUpdateSentenceDates,
+        stubMatchLicenceEvents: licence.stubMatchLicenceEvents,
+        stubApproveVariation: licence.stubApproveVariation,
+        stubReferVariation: licence.stubReferVariation,
+        stubUpdateStandardConditions: licence.stubUpdateStandardConditions,
+        stubGetHdcLicencesForOffender: licence.stubGetHdcLicencesForOffender,
+        stubGetVariationsSubmittedByRegionForOffender: licence.stubGetVariationsSubmittedByRegionForOffender,
+        stubGetLicencePolicyConditions: licence.stubGetLicencePolicyConditions,
+        stubGetActivePolicyConditions: licence.stubGetActivePolicyConditions,
+        stubGetPolicyChanges: licence.stubGetPolicyChanges,
+
+        stubGetPduHeads: community.stubGetPduHeads,
+        stubGetStaffDetails: community.stubGetStaffDetails,
+        stubGetStaffDetailsByStaffId: community.stubGetStaffDetailsByStaffId,
+        stubGetStaffDetailsByStaffCode: community.stubGetStaffDetailsByStaffCode,
+        stubGetStaffDetailsByList: community.stubGetStaffDetailsByList,
+        stubGetManagedOffenders: community.stubGetManagedOffenders,
+        stubGetAnOffendersManagers: community.stubGetAnOffendersManagers,
+        stubGetUserDetailsByUsername: community.stubGetUserDetailsByUsername,
+        stubAssignRole: community.stubAssignRole,
+
+        searchPrisonersByNomisIds: prisonerSearch.searchPrisonersByNomisIds,
+        searchPrisonersByBookingIds: prisonerSearch.searchPrisonersByBookingIds,
+        searchPrisonersByReleaseDate: prisonerSearch.searchPrisonersByReleaseDate,
+
+        stubGetPrisonUserDetails: prison.stubGetUserDetails,
+        stubGetPrisonUserCaseloads: prison.stubGetUserCaseloads,
+        stubGetPrisonerDetail: prison.stubGetPrisonerDetail,
+        stubGetPrisonInformation: prison.stubGetPrisonInformation,
+        stubGetHdcStatus: prison.stubGetHdcStatus,
+        stubGetPrisons: prison.stubGetPrisons,
+
+        stubGetProbationer: probationSearch.stubGetProbationer,
+        stubGetOffendersByCrn: probationSearch.stubGetOffendersByCrn,
+        stubGetOffendersByNomsNumber: probationSearch.stubGetOffendersByNomsNumber,
+
+        sendDomainEvent: events.sendDomainEvent,
+        sendPrisonEvent: events.sendPrisonEvent,
+        sendProbationEvent: events.sendProbationEvent,
+        purgeQueues: events.purgeQueues,
+        getNoneWorkingDays: (): Promise<Moment> =>
+          new UkBankHolidayFeedService().getEnglishAndWelshHolidays().then(ukHolidays => {
+            const appointmentDate = moment().add(1, 'year').add(1, 'week').day(7)
+            while (isBankHolidayOrWeekend(appointmentDate, ukHolidays)) {
+              appointmentDate.add(1, 'day')
+            }
+            return appointmentDate
+          }),
+      })
+    },
+    baseUrl: 'http://localhost:3007',
+    excludeSpecPattern: '**/!(*.cy).ts',
+    specPattern: 'integration_tests/integration/**/*.cy.{js,jsx,ts,tsx}',
+    supportFile: 'integration_tests/support/index.ts',
+  },
+})
