--- cypress.json
@@ -1,19 +1,0 @@
-{
-  "chromeWebSecurity": false,
-  "baseUrl": "http://localhost:3007",
-  "ignoreTestFiles": "**/!(*.spec).ts",
-  "fixturesFolder": "integration_tests/fixtures",
-  "integrationFolder": "integration_tests/integration",
-  "pluginsFile": "integration_tests/plugins/index.ts",
-  "screenshotsFolder": "integration_tests/screenshots",
-  "supportFile": "integration_tests/support/index.ts",
-  "videosFolder": "integration_tests/videos",
-  "reporter": "cypress-multi-reporters",
-  "reporterOptions": {
-    "configFile": "reporter-config.json"
-  },
-  "videoUploadOnPasses": false,
-  "taskTimeout": 60000,
-  "viewportHeight": 1200,
-  "viewportWidth": 1300
-}
