+++ integration_tests/integration/approveLicence.cy.ts
