+++ integration_tests/integration/approveVariation.cy.ts
