+++ integration_tests/integration/rejectLicence.cy.ts
