+++ integration_tests/integration/viewAndPrintLicence.cy.ts
