+++ package.json
@@ -169,8 +169,8 @@
     "better-npm-audit": "^3.7.3",
     "cheerio": "^1.0.0-rc.10",
     "concurrently": "^6.5.1",
-    "cypress": "^8.7.0",
-    "cypress-axe": "^0.14.0",
+    "cypress": "^12.2.0",
+    "cypress-axe": "^1.2.0",
     "cypress-multi-reporters": "^1.5.0",
     "dotenv": "^16.0.3",
     "eslint": "^7.32.0",
